Branchcp

+++ src/main/java/com/company/Summative1CristieJBNicholas/controller/GameController.java
@@ -1,13 +1,16 @@
 package com.company.Summative1CristieJBNicholas.controller;
 
 
+import com.company.Summative1CristieJBNicholas.exception.ProductNotFoundException;
 import com.company.Summative1CristieJBNicholas.models.Games;
+
 import com.company.Summative1CristieJBNicholas.repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 public class GameController {
@@ -21,6 +24,20 @@ public class GameController {
         return repo.findAll();
     }
 
+    // get games by ID
+    @GetMapping("/games/{id}")
+    public Games getGameById(@PathVariable Integer id) {
+        // from echo-range-service class work
+        if (id < 1) {
+            throw new IllegalArgumentException("Game ID must be at least 1");
+        }
+        Optional<Games> returnVal = repo.findById(id);
+        if (returnVal.isPresent()){
+            return returnVal.get();
+        } else {
+            throw new ProductNotFoundException("No such console. id:  " + id);
+        }
+    }
     @GetMapping(value="/games/title/{title}")
     @ResponseStatus(HttpStatus.OK)
     public List<Games> getGamesbyTitle(@PathVariable String title)
@@ -52,6 +69,9 @@ public class GameController {
     @DeleteMapping(value="/games/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteGame(@PathVariable Integer id){
+        if (id < 1) {
+            throw new IllegalArgumentException("Cannot delete a Game unless Game ID is at least 1");
+        }
         repo.deleteById(id);
     }
